Remove duplicates: move funcs to utils, decorator for pool
