+++ ocsmesh/hfun/mesh.py
@@ -132,6 +132,7 @@ class HfunMesh(BaseHfun):
                 inter for ply in multipolygon for inter in ply.interiors]
 
             features = MultiLineString([*exteriors, *interiors])
+            # pylint: disable=E1123
             self.add_feature(
                 feature=features,
                 expansion_rate=expansion_rate,
@@ -161,39 +162,8 @@ class HfunMesh(BaseHfun):
 
         self.mesh.msh_t.value = values
 
+    @utils.requires_pool
     def add_feature(
-            self,
-            feature: Union[LineString, MultiLineString],
-            expansion_rate: float,
-            target_size: float = None,
-            nprocs=None,
-            max_verts=200,
-            proc_pool=None
-    ):
-        if proc_pool is not None:
-            self._add_feature_internal(
-                feature=feature,
-                expansion_rate=expansion_rate,
-                target_size=target_size,
-                pool=proc_pool,
-                max_verts=max_verts)
-
-        else:
-            # Check nprocs
-            nprocs = -1 if nprocs is None else nprocs
-            nprocs = cpu_count() if nprocs == -1 else nprocs
-            _logger.debug(f'Using nprocs={nprocs}')
-
-            with Pool(processes=nprocs) as pool:
-                self._add_feature_internal(
-                    feature=feature,
-                    expansion_rate=expansion_rate,
-                    target_size=target_size,
-                    pool=pool,
-                    max_verts=max_verts)
-            pool.join()
-
-    def _add_feature_internal(
             self,
             feature: Union[LineString, MultiLineString],
             expansion_rate: float,
@@ -229,7 +199,7 @@ class HfunMesh(BaseHfun):
         _logger.info('Repartitioning features...')
         start = time()
         res = pool.starmap(
-            repartition_features,
+            utils.repartition_features,
             [(linestring, max_verts) for linestring in feature]
             )
         feature = functools.reduce(operator.iconcat, res, [])
@@ -256,7 +226,7 @@ class HfunMesh(BaseHfun):
                     f"Transform apply took {time() - start2:f}")
 
         transformed_features = pool.starmap(
-            transform_linestring,
+            utils.transform_linestring,
             [(linestring, target_size) for linestring in feature]
         )
         _logger.info(f'Resampling features took {time()-start}.')
@@ -324,36 +294,3 @@ class HfunMesh(BaseHfun):
 
     def get_bbox(self, **kwargs):
         return self.mesh.get_bbox(**kwargs)
-
-
-def transform_linestring(
-    linestring: LineString,
-    target_size: float,
-):
-    distances = [0.]
-    while distances[-1] + target_size < linestring.length:
-        distances.append(distances[-1] + target_size)
-    distances.append(linestring.length)
-    linestring = LineString([
-        linestring.interpolate(distance)
-        for distance in distances
-        ])
-    return linestring
-
-
-def repartition_features(linestring, max_verts):
-    features = []
-    if len(linestring.coords) > max_verts:
-        new_feat = []
-        for segment in list(map(LineString, zip(
-                linestring.coords[:-1],
-                linestring.coords[1:]))):
-            new_feat.append(segment)
-            if len(new_feat) == max_verts - 1:
-                features.append(ops.linemerge(new_feat))
-                new_feat = []
-        if len(new_feat) != 0:
-            features.append(ops.linemerge(new_feat))
-    else:
-        features.append(linestring)
-    return features
